* Fixed pine trees

+++ src/com/github/begla/blockmania/Chunk.java
@@ -451,7 +451,7 @@ public final class Chunk extends RenderableObject implements Comparable<Chunk> {
         float offsetZ = _position.z * Configuration.CHUNK_DIMENSIONS.z;
 
         boolean drawFront, drawBack, drawLeft, drawRight, drawTop, drawBottom;
-        int blockToCheck = _parent.getBlock(getBlockWorldPosX(x), getBlockWorldPosY(y + 1), getBlockWorldPosZ(z));
+        byte blockToCheck = _parent.getBlock(getBlockWorldPosX(x), getBlockWorldPosY(y + 1), getBlockWorldPosZ(z));
 
         drawTop = isSideVisibleForBlockTypes(blockToCheck, block);
 
@@ -964,7 +964,7 @@ public final class Chunk extends RenderableObject implements Comparable<Chunk> {
      *
      * NOTE: Air has to be handled separatly. Otherwise the water surface would not be displayed due to the tessellation process.
      */
-    private boolean isSideVisibleForBlockTypes(int blockToCheck, int currentBlock) {
+    private boolean isSideVisibleForBlockTypes(byte blockToCheck, byte currentBlock) {
         return blockToCheck == 0x0 || blockToCheck == 0x6 || (Block.getBlockForType(blockToCheck).isBlockTypeTranslucent() && !Block.getBlockForType(currentBlock).isBlockTypeTranslucent());
     }
